requirements.txt
apex==0.9.10dev
facenet_pytorch==2.5.2
ipython==8.4.0
librosa==0.9.1
mmcv==1.5.2
moviepy==1.0.3
numpy==1.19.5
opencv_python==4.5.2.52
packaging==21.3
pandas==1.3.4
Pillow==9.1.1
pydub==0.25.1
scikit_image==0.18.2
skimage==0.0
torch==1.11.0
torchaudio==0.11.0
transformers==4.21.0.dev0
